BookImporter and rake book:import
With all the various factories in place, the BookImporter loops over
a collection of lessons and creates topics and exercises. After every-
thing is created, it then creates the prerequisites based on the topic
codes provided in the required background sections of each lesson. The
entry point into this import process is the rake task, book:import.