Topic prerequisites/subsequents
In modeling how a topic relates to other topics, I decided that the
simplest approach would be for a topic to keep track of all its
prerequisite and subsequent topics. This makes reading very fast but
adds some complexity when writing. With this schema, it's not possible
to insert a prerequiste topic without traveling down the tree and
updating all subsequent topics. It's also not possible to easily build
a tree representation of all topics. However, since each lesson in the
book references all its prerequisites at once, this should get us
started.