Exercises#show renders with topic layout
I wanted the URL for exercises to be nested under topics. For example:

/topics/123/part-1

However, the number 1 in the URL is the part number or order of the
exercise, so I had to extract that from the Book::Lesson in the
ExerciseFactory. That was the hard part.

The easy part was rendering the topics layout of exercises#show.