MaterialFactory
This class uses the LessonPartSplitter to grab each line of text from
the materials fields of a given lesson part (our Exercise). It returns
an array of arrays, which is confusing, but simplistic. The idea is
that the first array is an index of parts (e.g., 0 == Part 1). The
inner array are the materials for a given exercise.