Removing not-null constraint on exercises name
As it turns out not every exercise is divided into parts. I didn't
notice this earlier because the ExerciseFactory conveniently scanned
the book_lesson.materials_and_procedures looking for something like
"Part 1." If the text wasn't divided up that way, it would loop over
an empty array. In other words, it would skip that exercise. Not good.
Now exercise names are optional, which lets topics with only one
(unnamed) exercise be saved. Rock!