Reworking MaterialFactory
Materials aren't always broken into parts but exercises require a
part. Therefore you will always ask for materials by part, even if
just part #1. The MaterialFactory now keeps track of materials that
are grouped by parts and materials that are not. Materials that are
not get associated to all exercises. At least that's the plan. I added
a few more test cases but we won't really know until we see it in a
browser.