Book import won't recreate models that already exist
When importing materials, it's important not to recreate materials
that are already created for a given exercise. This is due to the fact
that people may have edited that record in production and running a
new import would undo that work. But in order for materials to be
found for an exercise, exercises must also not be deleted when running
an import. Same thing goes for topics. So now all factories
find_or_initialize_by respective fields.

Along with this change, I reorganized the factory tests.