ApplicationController gets require_user and require_admin
We now have two methods of restricting access to a controller action:
require_user and require_admin. Both these methods rely on a user_id
to be in session. If the check fails, they redirect to the homepage
with an error message.

I implemented these checks on TopicsController and reworked the tests
for that controller to reflect the new behavior.