MessagesHelper.message_path
When linking to a message there could a few different paths depending
on what the message relates to. If it relates to a topic, it should
link to the topic with the message as an anchor. If it relates to an
exercise, it should link to the exercise with the message as an
anchor. And if it relates to another message, making it a reply, it
should link to that message's topic or exercise relationship. In the
last case, the anchor should still be the current message.