Markdown previews wherever there are textareas
Anywhere that people input text into a textarea will be Markdown
supported. This makes it easy for people to add basic formatting
without writing HTML or implementing a WYSIWYG editor.

A key part of supporting Markdown is to let people preview their text
before they submit it. I built a MarkdownPreview JS module that auto-
matically adds a "Preview" before all textarea fields. Clicking the
button takes you to a preview page in a new tab.

Markdown is supported in the all with RedCarpet, which has native
extensions so rendering should be fast.

Closes #40, closes #35.