Message archiving
Now, next to the close button there's also an archive button. Clicking
it pops up a confirmation dialogue before hitting the destory action
on the MessagesController, which is admin-only.

I took the chance to simplify fixtures here. I also added two new
scopes to the Message model, archived and unarchived.

There is decidedly no "unarchive" feature. We decided that would be
a console-only task.