YahooGroup and rake yahoo_group:import
The YahooGroup class is instantiated from a cookie string, which you
can get by taking a peak in the web inspector of your favorite
browser. It downloads all the message IDs first from the undocumented
Yahoo API that powers there web app. Then it hits the API again for
each message to get the full details. These details are passed to the
MessageFactory and then the message is saved. All the messages are
currently invalid because they don't have all their fields.