User helper methods: #stripe_customer, #subscribed?
Because calling to Stripe requires an API call, we want to do as few
as possible. In one request, there's no need to hit Stripe more than
once to find a user's associated Stripe Customer. The #stripe_customer
method handles that. Then a helper for the various edge cases of a
subscription, #subscribed?, makes it easy to tell if a user is paying.